refactor (image) : 이미지 도메인 로직 완전 분리 및 유즈케이스 테스트용이하게 수정
